requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.5.2
attrs==23.1.0
Brotli==1.0.9
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.4
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
dj-database-url==2.1.0
Django==4.0.2
django-crum==0.7.9
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dulwich==0.21.5
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
Pillow==9.1.1
pkginfo==1.9.6
platformdirs==3.8.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pydantic==2.0.2
pydantic_core==2.1.2
pyproject_hooks==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
rapidfuzz==2.15.1
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.8.10
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.27g.0
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.7.6
typing_extensions==4.7.1
tzdata==2022.1
ua-parser==0.10.0
urllib3==1.26.16
user-agents==2.2.0
uvicorn==0.22.0
virtualenv==20.23.1
webencodings==0.5.1
whitenoise==6.5.0
zipp==3.16.1
